--- a/.gitignore
+++ b/.gitignore
@@ -20,6 +20,8 @@ server/.env.development.local
 server/.env.test.local
 server/.env.production.local
 
+docs/
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*