--- a/client/src/App.css
+++ b/client/src/App.css
@@ -92,7 +92,8 @@ body.dark .btn-config .MuiButtonBase-root { color: #e7e7e7; }
 body.dark ::-webkit-calendar-picker-indicator { filter: invert(1); }
 body.dark .main__button { box-shadow: 2px 1000px 1px #212121 inset; }
 body.dark .projects__viewMore:hover { background-color: #79c3ff; color: #212121; }
-body.dark .footer__socialBtn a:hover { background-color: #FFFFFF; color: #212121; }
+body.dark .footer__socialBtn a:hover svg { color: #212121 !important; }
+body.dark .footer__socialBtn a svg { color: rgb(66 133 244) !important }
 body.dark .main__divider { background-image: linear-gradient(315deg, #0cbaba 0%, #f400eb 74%); }
 body.dark .loader { background-color: rgba(114, 114, 114, 0.61); }
 body.dark .form__contact { box-shadow: 0 1px 2px 0 rgb(0 0 0 / 60%), 0 1px 3px 1px rgb(0 0 0 / 30%); }
@@ -359,11 +360,6 @@ label { user-select: none; }
 .btn__changeMode:hover { box-shadow: 0px 0px 11px #9c9b98; }
 .main {
     margin: auto 30px;
-    padding-top: 180px;
-    transition: all .4s;
-}
-.main__projects {
-    margin: auto 30px;
     padding-top: 100px;
     transition: all .4s;
 }
@@ -618,9 +614,9 @@ select {
 .contact__btn:hover { background-color: #FFFFFF; color: #212121; }
 .get_in_touch { text-align: center; }
 img {
-    width: 100% !important;
-    height: 350px !important;
+    max-width: 100%;
     object-fit: cover !important;
+    transition: all .4s;
 }
 .contact__form { padding: 15px; }
 .contact__infoField {
@@ -752,8 +748,8 @@ img {
 }
 .footer__socialBtn {
     display: grid;
-    grid-template-columns: auto auto auto auto;
-    width: 50%;
+    grid-template-columns: auto auto;
+    width: 100%;
     margin: auto;
 }
 .footer__socialBtn a {
@@ -762,7 +758,7 @@ img {
     display: inline-block;
     width: 40px;
     height: 40px;
-    color: #FFFFFF;
+    color: #1a73e8;
     text-align: center;
     transition: all .4s;
     padding-top: 10px;
@@ -771,9 +767,9 @@ img {
     border-radius: 100%;
     text-decoration: none;
     background-color: none;
-    border: 2px solid #FFFFFF;
+    border: 2px solid #1a73e8;
 }
-.footer__socialBtn a:hover { background-color: #FFFFFF; color: #1a73e8; }
+.footer__socialBtn a:hover { background-color: #1a73e8; color: #FFF; }
 .footer__copyright { margin-top: 40px; width: 100%; text-align: center; }
 .content__transition { display: none; }
 .content__fadeIn {