--- a/client/src/App.css
+++ b/client/src/App.css
@@ -81,7 +81,7 @@ body.dark .otp input {
 }
 body.dark .main__title,
 body.dark .projects__viewMore,
-body.dark .link { color: #79c3ff !important; }
+body.dark .link, body.dark .blue-text { color: #79c3ff !important; }
 body.dark .link:focus { color: #5cb5ff !important; }
 body.dark .otp input { border-color: rgb(59, 57, 57); }
 body.dark .navbar { box-shadow: 0 4px 5px 0 rgba(0,0,0,0.14),0 1px 10px 0 rgba(0,0,0,0.12),0 2px 4px -1px rgba(0,0,0,0.2); }
@@ -92,7 +92,7 @@ body.dark .btn-config .MuiButtonBase-root { color: #e7e7e7; }
 body.dark ::-webkit-calendar-picker-indicator { filter: invert(1); }
 body.dark .main__button { box-shadow: 2px 1000px 1px #212121 inset; }
 body.dark .projects__viewMore:hover { background-color: #79c3ff; color: #212121; }
-body.dark .footer__socialBtn a:hover svg { color: #212121 !important; }
+body.dark .footer__socialBtn a:hover svg { color: #FFF !important; }
 body.dark .footer__socialBtn a svg { color: rgb(66 133 244) !important }
 body.dark .main__divider { background-image: linear-gradient(315deg, #0cbaba 0%, #f400eb 74%); }
 body.dark .loader { background-color: rgba(114, 114, 114, 0.61); }
@@ -114,6 +114,10 @@ body.dark .MuiInput-underline:hover:not(.Mui-disabled):before { border-bottom: 1
 .MuiSvgIcon-root { color: #212121; }
 button, a { user-select: none; cursor: pointer; }
 label { user-select: none; }
+.monospace {
+    font-weight: lighter;
+    font-family: 'Roboto Mono', monospace;
+}
 .privacy-policy {
     font-family: 'Noto Sans JP', sans-serif;
     margin: auto 150px;
@@ -185,6 +189,7 @@ label { user-select: none; }
 .m-5 { margin: 5px !important; }
 .m-10 { margin: 10px !important; }
 .p-12 { padding: 12px !important; }
+.pt-180 { padding-top: 180px !important; }
 .ml-10 { margin-left: 10px !important; }
 .mr-10 { margin-right: 10px !important; }
 .mt-10 { margin-top: 10px !important; }
@@ -614,7 +619,7 @@ select {
 .contact__btn:hover { background-color: #FFFFFF; color: #212121; }
 .get_in_touch { text-align: center; }
 img {
-    max-width: 100%;
+    max-width: 100% !important;
     object-fit: cover !important;
     transition: all .4s;
 }
@@ -681,6 +686,7 @@ img {
     width: 0;
     z-index: -1;
 }
+.blue-text { color: #1a73e8; }
 .notification-success { background-color: #55b559; }
 .notification-warning { background-color: #ffcc00; color: black; }
 .notification-danger { background-color: #f55145; }
@@ -893,6 +899,8 @@ blockquote em {font-style: italic;}
     outline: 0;
     transition: color 0.2s ease, background-color 0.2s ease, border-color 0.2s ease, box-shadow 0.2s ease;
 }
+.large { font-size: 60px; }
+.center-object { margin: auto; max-width: 70%; }
 .justify-center { justify-content: center; }
 @keyframes octocat-wave {
     0%,100%{transform:rotate(0)}