--- a/client/src/App.css
+++ b/client/src/App.css
@@ -194,6 +194,8 @@ label { user-select: none; }
 .m-5 { margin: 5px !important; }
 .m-10 { margin: 10px !important; }
 .p-12 { padding: 12px !important; }
+.pl-10 { padding-left: 10px !important; }
+.pr-10 { padding-right: 10px !important; }
 .p-5p { padding: 5% !important; }
 .pt-180 { padding-top: 180px !important; }
 .ml-10 { margin-left: 10px !important; }
@@ -614,7 +616,7 @@ select {
 }
 .contact__container {
     display: grid;
-    grid-template-columns: auto auto;
+    grid-template-columns: 50% 50%;
     padding: 10px;
 }
 .contact__body, .contact__me { padding: 15px; }
@@ -1020,6 +1022,7 @@ blockquote em {font-style: italic;}
     .center-object { max-width: 100%; }
 }
 @media only screen and (max-width: 600px) {
+    .contact__container p { padding: 0px !important; }
     .contact__container { grid-template-columns: auto; }
     .animation__message { width: 100%; height: auto; }
     .btn__outline { width: 65%; }