--- a/client/src/App.css
+++ b/client/src/App.css
@@ -1,4 +1,4 @@
-/*! Todo Application - CSS Version 0.5.3 Copyright 2021 */
+/*! Otlio - CSS Version 0.5.4 Copyright 2021 */
 @import url('https://fonts.googleapis.com/css2?family=Chilanka&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Raleway&display=swap');