--- a/client/src/App.js
+++ b/client/src/App.js
@@ -61,8 +61,8 @@ export default function App() {
     .catch(err => {
       setUserData({ type: {}, credentials: {}, security: {}, ...err.response.data, isLoading: false, authenticated: false })
       localStorage.setItem('XSRF-TOKEN', err.response.data['XSRF-TOKEN'])
-      if(err.response.status === 302 && err.response.data.type.mfa && (window.location.pathname !== '/login' && window.location.pathname !== '/logout')) window.location='/login'
-      if(err.response.status === 302 && err.response.data.type.verifyAccount && (window.location.pathname !== '/get-started' && window.location.pathname !== '/logout' && window.location.pathname.split('/')[1] !== 'verify' )) window.location='/get-started'
+      if(err.response.status === 302 && err.response.data.type.mfa && (window.location.pathname !== '/login' && window.location.pathname !== '/logout' && window.location.pathname !== '/support')) window.location='/login'
+      if(err.response.status === 302 && err.response.data.type.verifyAccount && (window.location.pathname !== '/get-started' && window.location.pathname !== '/logout' && window.location.pathname !== '/support' && window.location.pathname.split('/')[1] !== 'verify' )) window.location='/get-started'
       if(err.response.data.message && err.response.data.message !== "No auth token") setNotification(NOTIFICATION_TYPES.DANGER, err.response.data.message)
     })
     console.log("%c%s","color: red; background: yellow; font-size: 24px","WARNING!")