--- a/client/src/App.js
+++ b/client/src/App.js
@@ -1,6 +1,6 @@
+import axios from 'axios'
 import React, { useState, useEffect } from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
-import axios from 'axios'
 
 import { setNotification, NOTIFICATION_TYPES } from './libraries/setNotification'
 import "./App.min.css"