--- a/client/src/components/404.component.js
+++ b/client/src/components/404.component.js
@@ -3,23 +3,26 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEnvelope } from '@fortawesome/free-solid-svg-icons'
 import { faGithub } from '@fortawesome/free-brands-svg-icons'
 import Image from '../img/4f4adcbf8c6f66dcfc8a3282ac2bf10a.webp'
+import { Tooltip } from '@material-ui/core'
 
-const Landing = ({ userData }) => {
+const PageNotFound = ({ userData }) => {
     return (
         <div className="main">
             <div className="contact__container">
-                <img src={Image} className=""/>
-                <div className="center-vertical p-50">
-                    <h1>404 - Page Not Found</h1>
-                    <h2 className="mt-20">We're sorry, we couldn't find the page you requested.</h2>
-                    <button className="oauth-box google isCentered block mt-30 mb-20 p-12 button" onClick={() => userData.authenticated ? window.location='/' : window.location='/welcome'}>Back</button>
-                    <div class="footer__socialBtn">
-                        <a href="mailto:stanleyowen06@gmail.com">
+                <img src={Image} className="center-object"/>
+                <div className="center-object">
+                    <h1 className="blue-text monospace large">404</h1>
+                    <h1 className="blue-text monospace">Page Not Found</h1>
+                    <h3 className="mt-20 monospace">Oops! We are sorry, we can't find the page you were looking for.</h3>
+                    <h3 className="mt-20 monospace">If you're experiencing a critical issue, please <a className="animation__underline" href="mailto:stanleyowen06@gmail.com">email support</a>.</h3>
+                    <button className="oauth-box google isCentered block mt-30 mb-20 p-12 button" onClick={() => userData.authenticated ? window.location='/' : window.location='/welcome'}>Back to Home</button>
+                    <div class="footer__socialBtn mb-20">
+                        <Tooltip title="Email Support" arrow><a href="mailto:stanleyowen06@gmail.com">
                             <FontAwesomeIcon icon={faEnvelope} style={{ fontSize: "1.8em" }} />
-                        </a>
-                        <a href="https://github.com/stanleyowen">
+                        </a></Tooltip>
+                        <Tooltip title="View Code on GitHub" arrow><a href="https://github.com/stanleyowen/todo-application/">
                             <FontAwesomeIcon icon={faGithub} style={{ fontSize: "1.8em" }} />
-                        </a>
+                        </a></Tooltip>
                     </div>
                 </div>
             </div>
@@ -27,4 +30,4 @@ const Landing = ({ userData }) => {
     )
 }
 
-export default Landing
+export default PageNotFound