… CONTRIBUTING.md
… README.md
… client/package.json
… client/public/favicon.ico
… client/public/index.html
… client/public/logo192.png
… client/public/logo512.png
… client/public/manifest.json
… client/src/App.min.css
… client/src/components/account.component.js
… client/src/components/edit.component.js
… client/src/components/forgot-password.component.js
… client/src/components/login.component.js
… client/src/components/logout.component.js
… client/src/components/navbar.component.js
… client/src/components/privacy-policy.component.js
… client/src/components/register-oauth.component.js
… client/src/components/register.component.js
… client/src/components/req-oauth.component.js
… client/src/components/reset-password.component.js
… client/src/components/support.component.js
… client/src/components/terms-and-condition.component.js
… client/src/components/verify-account.component.js
… client/src/components/welcome.component.js
… client/src/img/2405976c7c1485050e78fcf54ca4bfe4.webp
… client/src/img/4f4adcbf8c6f66dcfc8a3282ac2bf10a.webp
… client/src/img/765cbf066c6e4c42444a0ce9c2fb7949.webp
… client/src/img/7c27535f88bae9519ceb14a8983c57ff.webp
… client/src/libraries/setNotification.js
… client/src/libraries/validation.js