**What's Changed in v0.5.3:**
1. Introducing **Otlio** (bb4ca3a3acf9b9d8cb8d770e39f7ebdfaaaf0898)
2. Redesigned Otlio Icon (#167) (f3ef5f788dc16cf90857af44b06283c7da9381b3)
3. Added `Semgrep` CI (f2e8646dc13df7e0cfb996d40b17db82a3158ae2)
4. Added Support Feature for All Accounts (#174) (67b6665731b459accd6e757bcf80bf739c2d52c9) (8d4eda0db36ddff01a00c7068b2e94231e1b7c67) (c25b5b53109291e4987ff77a6b957d51c926d328) (ea848af9829a1eb7b3957700a273132d9aab1f70)
5. Added 404 Error Page (#169) (3d0b46e75a4833960d7261c9300352523f7ded3a) (b8b8a2fbd399de14e58703863fc6ebe7f5d1cbfe)
6. Added Drag and Drop Feature (#172) (e544f541806a02745f828de8d7a7db6f2c3a9aa6)
7. Added `cross-env` Dependency (1438fb8b9ddda3fdc22f509b409caa8580c79c2a):
8. Added `react-beautiful-dnd` Dependency (f615b006f1f48a4570b914e5f005aeb08344f573)
9. Preview Image from `CDN` Server for Speed and Optimization (c25b5b53109291e4987ff77a6b957d51c926d328)
10. Improved Accessibility (d8b454465e067fa0a70b9b50777e0f6e1359c0a6)
11. Improved User Interface (5093fd2497a130a977f350bc7f99e9bc03bb299c) (c636a4cbdf5bf7d42d293e1826eda3a6631731a8):
    * `Modal` (fec93e4210b2229ebd1a4b29fa60184f7241a2b0)
    * Welcome Component (#173) (759a6acf7ae1deea2d6b779e6f6d6cb525747e84) (ad6e71e548b65f815e9bf3499838337a813cda85) (8859b38c873c8a0b7a3fe795d8c71823c4bddb67) (066e744edeaefd9c4faabc7fe0a20a8a50e6ee56) (f957304e49b78d31fd950c39751ebfdcac54a152)
    * Account Settings (85dd38baddfb1659becee9ffbbe309c8577f5349)
12. Improved Performance and Code Base Quality (ea34d1abdd0e3d7a3edd77ca31178a223226bef5) (dd912bd5994512d3415e0434b1b6b41675ee446a) (c636a4cbdf5bf7d42d293e1826eda3a6631731a8) (dea3b73d8dfc28f8efccfadb28f5c9aa28cf99f1) (9be186fc2437787a7037612d19c6bbd4c5ffe153)
13. Improved Security for:
    * Parsing Data in Support Feature (#175) (45023859aa76d30daee2a03291fcd38b03fceb70)
    * Link `Attribute` (743f38a1a57a4c6ab0abcc7a95bbd36a0c6049ef)
    * Displaying data from `server` using `innerText` instead of `innerHTML` (#182) (579b032cb578ed8cc2620a19e9da0d54674bb238)
14. Optimized Server `API` (f636ce0111f5259c7cdcfc6e1920fca23eb8163c)
15. Update and Upgrade `Dependencies` to the Latest Version (6a45ad3101ac10f3946b40331e8dc7af43e16e75) (1438fb8b9ddda3fdc22f509b409caa8580c79c2a):
    * `dotenv` dependency to `10.0.0`
    * `nodemailer` dependency to `6.6.1`
    * `mongoose` dependency to `5.12.11`
16. Update `README.md` (4eaedf5ad32f180142e599595d58acc52aca64b9)
17. Resolved Issues of:
    * More than 10 codes are displayed after clicked Regenerate Code `button` (#170) (64cdc404dcad95485d7d75182acf3c5ab5e156a0)
    * 404 Error Page is Showing in Any Path (#171) (b8b8a2fbd399de14e58703863fc6ebe7f5d1cbfe)
    * Missing `return` statement (#176) (c6fc5f2b8cf0676f1d7ab3abbefd7c2492292ce5)
    * Vulnerable against `XSS` Attack (#177) (25ac473adf37811820b7af55c514211e0b1dea09)
    * Unresponsive Button in Backup Codes Modal (#179) (a3de7af63aabc549dd9c27e72a7c245cf2099779)
    * Date not being parser appropriately in the first and last of the month (#180) (b26ee520df8fdcb4cb2529a81c7495dff61ebe0b)
    * Not returning version number in `production` mode (#181) (03aaeb27164c28e37bcd7c98f856aa8c6501ca14) (1438fb8b9ddda3fdc22f509b409caa8580c79c2a)

Details About Changes in v0.5.2:
| Client | `build\static\js\2.6ca8d93c.chunk.js` | `build\static\js\main.2591d28a.chunk.js` | `build\static\css\main.030489ff.chunk.css` | `build\static\js\3.668acfa5.chunk.js` | `build\static\js\runtime-main.3cbd5898.js`
------ | --------------------------------------- | ------------------------------------------- | ---------------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
| v0.5.4 | 188.6 kB (+35.01 kB) | 32.1 kB (+2.86 kB) | 5.4 kB (+286 B) | 1.62 KB (-2 B) | 1.16 KB (-3 B) |
| v0.5.3 | 153.59 kB (+979 B) | 29.24 kB (-807 B) | 5.12 kB (+382 B) | 1.62 kB (+1 B) | 1.17 KB (+2 B) |
| v0.5.2 | 152.63 kB (+200 B) | 30.03 kB (+220 B) | 4.75 kB (+20 B) | 1.62 kB (+60 B) | 1.16 kB |
| v0.5.1 | 152.43 kB (+165 B) | 29.81 kB (+1.47 kB) | 4.73 kB (+193 B) | 1.56 kB | 1.16 kB |
| v0.5.0 | 152.27 kB (+3.16 kB) | 28.34 kB (+428 B) | 4.54 kB (+37 B) | 1.56 kB | 1.16 kB |

| Server | Size |
| ------- | ---- |
| v0.5.4 | 200.81 KB (-12.35 kB) |
| v0.5.3 | 213.16 KB (-3.16 kB) |
| v0.5.2 | 216.32 kB (-2.86 kB) |
| v0.5.1 | 219.18 kB (+30 B) |
| v0.5.0 | 219.15 kB (+7.2 kB) |

For more information about changes in this release, please visit our [project roadmaps v0.5.4](https://github.com/stanleyowen/todo-application/projects/34)